Extract some utils from main Postgres adapter file
